// src/BookBarter.React/src/api/generated/models/genre-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BookBarter.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface GenreDto
 */
export interface GenreDto {
    /**
     * 
     * @type {number}
     * @memberof GenreDto
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof GenreDto
     */
    'name'?: string | null;
}