// src/BookBarter.React/src/api/generated/models/create-book-command.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BookBarter.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface CreateBookCommand
 */
export interface CreateBookCommand {
  /**
   *
   * @type {string}
   * @memberof CreateBookCommand
   */
  isbn?: string | null;
  /**
   *
   * @type {string}
   * @memberof CreateBookCommand
   */
  title?: string | null;
  /**
   *
   * @type {string}
   * @memberof CreateBookCommand
   */
  publicationDate?: string;
  /**
   *
   * @type {Array<number>}
   * @memberof CreateBookCommand
   */
  authorsIds?: Array<number> | null;
  /**
   *
   * @type {number}
   * @memberof CreateBookCommand
   */
  genreId?: number;
  /**
   *
   * @type {number}
   * @memberof CreateBookCommand
   */
  publisherId?: number;
}